change foreign key types

+++ database/migrations/2023_08_19_212913_component_author_description_and_tags.php
@@ -19,8 +19,8 @@ return new class extends Migration
 
         Schema::create('component_tag', function (Blueprint $table) {
             $table->id();
-            $table->integer('component_id');
-            $table->integer('tag_id');
+            $table->bigInteger('component_id');
+            $table->bigInteger('tag_id');
             $table->timestamps();
 
             $table
